requirements.txt
docopt==0.6.2
flake8==3.5.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
stomp.py==4.1.21
